refactor Requests class to separate file
